Update key.h
add interruption for the key, and save the language setting. now you can press key0 when reading to pause the reading or press it when finish the story to change the language. an internal LED will lid up for English and turn off for Chinese on start or on change. the change of language will apply to the next play. you can change the story by reset